requirements.txt
requests
beautifulsoup4
lxml
fastapi
uvicorn
google-generativeai
python-dotenv
